Towards deploying AE training. Here: Use of augmented data.

+++ 4_ICCS/trial_1.py
@@ -60,25 +60,27 @@ def train_AE(loader, model, optimizer, criterion, scaler, model_identifier, curr
     for _batch_idx, (_data_0, _targ_0) in enumerate(loader):
         t, c, h, d, w = _data_0.shape
         _data_u_x = torch.reshape(
-            _data_0[:, 0, :, :, :].float(), (t, 1, h, d, w)).to(device=device)
+            _data_0[:, 0, :, :, :], (t, 1, h, d, w)).to(device=device)
         _data_u_y = torch.reshape(
-            _data_0[:, 1, :, :, :].float(), (t, 1, h, d, w)).to(device=device)
+            _data_0[:, 1, :, :, :], (t, 1, h, d, w)).to(device=device)
         _data_u_z = torch.reshape(
-            _data_0[:, 2, :, :, :].float(), (t, 1, h, d, w)).to(device=device)
+            _data_0[:, 2, :, :, :], (t, 1, h, d, w)).to(device=device)
         _targ_u_x = torch.reshape(
-            _targ_0[:, 0, :, :, :].float(), (t, 1, h, d, w)).to(device=device)
+            _targ_0[:, 0, :, :, :], (t, 1, h, d, w)).to(device=device)
         _targ_u_y = torch.reshape(
-            _targ_0[:, 1, :, :, :].float(), (t, 1, h, d, w)).to(device=device)
+            _targ_0[:, 1, :, :, :], (t, 1, h, d, w)).to(device=device)
         _targ_u_z = torch.reshape(
-            _targ_0[:, 2, :, :, :].float(), (t, 1, h, d, w)).to(device=device)
+            _targ_0[:, 2, :, :, :], (t, 1, h, d, w)).to(device=device)
 
         _data_1 = torch.cat((_data_u_y, _data_u_z, _data_u_x), 1).to(device)
         _data_2 = torch.cat((_data_u_z, _data_u_x, _data_u_y), 1).to(device)
         _targ_1 = torch.cat((_targ_u_y, _targ_u_z, _targ_u_x), 1).to(device)
         _targ_2 = torch.cat((_targ_u_z, _targ_u_x, _targ_u_y), 1).to(device)
 
-        _data = torch.cat((_data_0.to(device), _data_1.to(device), _data_2.to(device)), 0).to(device)
-        _targ = torch.cat((_targ_0.to(device), _targ_1.to(device), _targ_2.to(device)), 0).to(device)
+        _data = torch.cat((_data_0.to(device), _data_1.to(
+            device), _data_2.to(device)), 0).to(device)
+        _targ = torch.cat((_targ_0.to(device), _targ_1.to(
+            device), _targ_2.to(device)), 0).float().to(device)
 
         with torch.cuda.amp.autocast():
             _pred = model(_data).float().to(device)
